requirements.txt
python-dotenv
Flask
Flask-PyMongo
pymongo[srv]